Deleted seasonal summary work to pull down Macks version
