+++ data/dummy_data/dummy_data_game_ts.csv
@@ -1,6 +1,6 @@
 "game_id","season","type","date_time","away_team_id","home_team_id","away_goals","home_goals","outcome","home_rink_side_start","venue","venue_link","venue_time_zone_id","venue_time_zone_offset","venue_time_zone_tz"
 "2012030221","20122013","P",2013-05-16,"3","6",2,3,"home win OT","left","TD Garden","/api/v1/venues/null","America/New_York",-4,"EDT"
-"2012030222","20122013","P",2013-05-19,"3","6",2,5,"home win REG","left","TD Garden","/api/v1/venues/null","America/New_York",-4,"EDT"
+"2012030222","20122013","R",2013-05-19,"3","6",2,5,"home win REG","left","TD Garden","/api/v1/venues/null","America/New_York",-4,"EDT"
 "2012030223","20132014","P",2013-05-21,"6","3",2,1,"away win REG","right","Madison Square Garden","/api/v1/venues/null","America/New_York",-4,"EDT"
 "2012030224","20132014","R",2013-05-23,"6","3",3,4,"home win OT","right","Madison Square Garden","/api/v1/venues/null","America/New_York",-4,"EDT"
 "2012030225","20132014","P",2013-05-25,"3","6",1,3,"home win REG","left","TD Garden","/api/v1/venues/null","America/New_York",-4,"EDT"
@@ -13,8 +13,8 @@
 "2012030233","20122013","R",2013-05-20,"16","17",1,3,"home win REG","right","Joe Louis Arena","/api/v1/venues/null","America/Detroit",-4,"EDT"
 "2012030234","20122013","P",2013-05-24,"16","17",0,2,"home win REG","right","Joe Louis Arena","/api/v1/venues/null","America/Detroit",-4,"EDT"
 "2012030235","20122013","P",2013-05-26,"17","16",1,4,"home win REG","right","United Center","/api/v1/venues/null","America/Chicago",-5,"CDT"
-"2012030236","20122013","P",2013-05-28,"16","17",4,3,"away win REG","right","Joe Louis Arena","/api/v1/venues/null","America/Detroit",-4,"EDT"
+"2012030236","20122013","R",2013-05-28,"16","17",4,3,"away win REG","right","Joe Louis Arena","/api/v1/venues/null","America/Detroit",-4,"EDT"
 "2012030237","20122013","P",2013-05-30,"17","16",1,2,"home win OT","right","United Center","/api/v1/venues/null","America/Chicago",-5,"CDT"
 "2012030121","20122013","P",2013-05-02,"9","8",4,2,"away win REG","right","Centre Bell","/api/v1/venues/null","America/Montreal",-4,"EDT"
-"2012030122","20122013","P",2013-05-03,"9","8",1,3,"home win REG","right","Centre Bell","/api/v1/venues/null","America/Montreal",-4,"EDT"
-"2012030123","20122013","P",2013-05-05,"8","9",1,6,"home win REG","left","Scotiabank Place","/api/v1/venues/null","America/New_York",-4,"EDT"
+"2012030122","20122013","R",2013-05-03,"9","8",1,3,"home win REG","right","Centre Bell","/api/v1/venues/null","America/Montreal",-4,"EDT"
+"2012030123","20122013","R",2013-05-05,"8","9",1,6,"home win REG","left","Scotiabank Place","/api/v1/venues/null","America/New_York",-4,"EDT"
